requirements.txt

pandas==2.0.3
numpy==1.25.0
statsmodels==0.14.0
scikit-learn==1.3.2
scipy==1.9.3
Flask==2.2.0
werkzeug==2.2.0
Flask-RESTful==0.3.9
python-dotenv==1.0.0
datetime==4.3
requests==2.31.0
pyjwt==2.7.0
